Scanning QR Fragment Working. Added PostScanFragment
