+++ SnapScan/app/build.gradle
@@ -38,20 +38,32 @@ dependencies {
     implementation 'com.google.android.material:material:1.8.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.core:core-splashscreen:1.0.0'
-    implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.5.1'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1'
+    implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.5.1'
+
+    // Navigation Dependencies
     implementation 'androidx.navigation:navigation-fragment:2.5.3'
     implementation 'androidx.navigation:navigation-ui:2.5.3'
-    implementation 'com.google.android.gms:play-services-maps:18.1.0'
-    implementation 'com.google.firebase:firebase-firestore:24.4.4'
+    // Testing Dependencies
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+    androidTestImplementation 'androidx.test:rules:1.5.0'
+
+    // QR scanner Dependencies
+    implementation 'com.journeyapps:zxing-android-embedded:4.3.0'
+    // Map Dependencies
+    implementation 'com.google.android.gms:play-services-maps:18.1.0'
     implementation 'com.google.android.gms:play-services-maps:18.1.0'
     implementation 'com.google.android.gms:play-services-location:21.0.1'
-    implementation 'com.journeyapps:zxing-android-embedded:4.3.0'
-    implementation platform('com.google.firebase:firebase-bom:31.2.3')
+
+    //Firebase Dependencies
     implementation 'com.google.firebase:firebase-analytics'
-    androidTestImplementation 'androidx.test:rules:1.5.0'
+    implementation 'com.google.firebase:firebase-firestore:24.4.4'
+    implementation platform('com.google.firebase:firebase-bom:31.2.3')
+
+    // Unique name Dependencies
+    implementation 'com.github.javafaker:javafaker:1.0.2'
+
 //    implementation 'com.google.android.gms:play-services-location:17.0.0'
 }
