+++ SnapScan/app/src/main/res/navigation/mobile_navigation.xml
@@ -13,7 +13,7 @@
 
     <fragment
         android:id="@+id/navigation_qr"
-        android:name="com.example.SnapScan.ui.qr.QRFragment"
+        android:name="com.example.SnapScan.ui.qr.QRScanFragment"
         android:label="@string/title_qr"
         tools:layout="@layout/fragment_qr" />
