+++ SnapScan/settings.gradle
@@ -11,6 +11,8 @@ dependencyResolutionManagement {
         google()
         mavenCentral()
     }
+
+
+    rootProject.name = "SnapScan"
+    include ':app'
 }
-rootProject.name = "SnapScan"
-include ':app'
